Use Capella GET endpoint to fetch user details
Currently, the SDK uses the /resolve endpoint on Capella to fetch
details for the current user such as the tresor ID or the tag encryption
key. This endpoint is only meant for translating between e-mails and
Gesundheitscloud/Zerokit user IDs and will shortly have its responses
changed to reflect this. Instead, the GET endpoint for the specific
user should be used for such requests. Note this requires the user to be
logged in.