Se/add share tresor (#74)
* Initilize adding shareTresor

* Share tresor fixed and works.

In previous version, tresorId was being used as a promise, not an actual
id. Also, share option is added to HealthCloud.js.

* Update shareTresor to work with email of grantee

In previous version, shareTresor was working with the zerokitId of the
grantee. A function is added in UserService which returns a user by
email address. Also, collectivizeTresor sends the granteeId to altair in the
current version, rather than
grantee's zerokitId.

* Resolve eslint errors

* Add initial unit tests

* Change collectivizeTresor to grantTresor

For better comprehension

* Upadte shareTresor not to be sketchy

Promise.all is added in shareTresor. Also, email is substituted with
alias, to be consistent with the rest of the sdk.

* Update tests, and fix shareTresor return value

"If shareTresor fails when zerokit shareTresor failed" added. Also,
return value of shareTresor used to be undefined and is fixed to
opertionId in this commit.

* Refactor "share" to "shareData" in HC sdk

Function "share" is  changed to "shareData" in HC sdk. Consequently,
"shareData" is changed to "grantTresor" in routes/userRoutes.

* Update sdk docs regarding shareData property

* Update test for grantTresor route failure

* Update test with userService resolveUserByAlias

Succeed and fail added for resolveUserByAlias in services.

* Correct userPass route test

* Remove log statements

* Update ZeroKitAdapter shareTresor by promise chain

ZeroKitAdapter shareTresor is reorganized with promise chain for cleaner
code.

* Refactor "shareTresor" and "grantTresor"

Method "shareTresor" in services' ZeroKitAdapter is changed to "grantPermission".
Method "grantTresor" in routes' userRoutes is changed to
"verifyShareAndGrantPermission". Refactoring is done in tests as well.

* Change shareData in healthcloud and fix test

For consistency of calling the internal functins as they are in
healthcloud as well, we change "shareData" to "grantPermission". However,
"shareData" might be a more comprehensible name and we might change it in
future. Also in test, "grantPermission" fails updated with "when" it fails.

* Update docs with grantPermission

"shareData" is changed to "grantPermission"